ai-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-generativeai
pandas
numpy
scikit-learn
requests
